Create FormHelper fake
