+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/verticalmode/VerticalModeInitialScreenFactory.kt
@@ -3,7 +3,7 @@ package com.stripe.android.paymentsheet.verticalmode
 import androidx.lifecycle.viewModelScope
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
 import com.stripe.android.paymentsheet.CustomerStateHolder
-import com.stripe.android.paymentsheet.FormHelper
+import com.stripe.android.paymentsheet.DefaultFormHelper
 import com.stripe.android.paymentsheet.LinkInlineHandler
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen
@@ -47,7 +47,7 @@ internal object VerticalModeInitialScreenFactory {
                 val paymentMethodCode = newPaymentSelection.paymentMethodCreateParams.typeCode
 
                 val linkInlineHandler = LinkInlineHandler.create(viewModel, viewModel.viewModelScope)
-                val formHelper = FormHelper.create(viewModel, linkInlineHandler, paymentMethodMetadata)
+                val formHelper = DefaultFormHelper.create(viewModel, linkInlineHandler, paymentMethodMetadata)
 
                 if (formHelper.requiresFormScreen(paymentMethodCode)) {
                     add(
